@@ -1,6 +1,8 @@
 package brigadier
 
-import "go.minekube.com/brigodier"
+import (
+	"go.minekube.com/brigodier"
+)
 
 var (
 	RegistryKeyArgument brigodier.ArgumentType = &RegistryKeyArgumentType{}
@@ -10,8 +12,13 @@ type RegistryKeyArgumentType struct {
 	Identifier string
 }
 
-func (r *RegistryKeyArgumentType) Parse(rd *brigodier.StringReader) (interface{}, error) {
+func (r *RegistryKeyArgumentType) Parse(rd *brigodier.StringReader) (any, error) {
 	return rd.ReadString()
 }
 
 func (r *RegistryKeyArgumentType) String() string { return "registry_key_argument" }
+
+type ByteArgumentType byte
+
+func (b ByteArgumentType) Parse(*brigodier.StringReader) (interface{}, error) { return byte(0), nil }
+func (b ByteArgumentType) String() string                                     { return "byte" }
