@@ -1,7 +1,7 @@
 package util
 
 // NBT is a named binary tag.
-type NBT map[string]interface{}
+type NBT map[string]any
 
 func (b NBT) Bool(name string) (bool, bool) {
 	val, ok := b.Uint8(name)
@@ -9,14 +9,14 @@ func (b NBT) Bool(name string) (bool, bool) {
 }
 
 func (b NBT) Int8(name string) (ret int8, ok bool) {
-	var val interface{}
+	var val any
 	if val, ok = b[name]; ok {
 		ret, ok = val.(int8)
 	}
 	return
 }
 func (b NBT) Uint8(name string) (ret uint8, ok bool) {
-	var val interface{}
+	var val any
 	if val, ok = b[name]; ok {
 		ret, ok = val.(uint8)
 	}
@@ -24,7 +24,7 @@ func (b NBT) Uint8(name string) (ret uint8, ok bool) {
 }
 
 func (b NBT) Int16(name string) (ret int16, ok bool) {
-	var val interface{}
+	var val any
 	if val, ok = b[name]; ok {
 		ret, ok = val.(int16)
 	}
@@ -32,7 +32,7 @@ func (b NBT) Int16(name string) (ret int16, ok bool) {
 }
 
 func (b NBT) Int32(name string) (ret int32, ok bool) {
-	var val interface{}
+	var val any
 	if val, ok = b[name]; ok {
 		ret, ok = val.(int32)
 	}
@@ -44,7 +44,7 @@ func (b NBT) Int(name string) (int, bool) {
 }
 
 func (b NBT) Int64(name string) (ret int64, ok bool) {
-	var val interface{}
+	var val any
 	if val, ok = b[name]; ok {
 		ret, ok = val.(int64)
 	}
@@ -52,7 +52,7 @@ func (b NBT) Int64(name string) (ret int64, ok bool) {
 }
 
 func (b NBT) Float32(name string) (ret float32, ok bool) {
-	var val interface{}
+	var val any
 	if val, ok = b[name]; ok {
 		ret, ok = val.(float32)
 	}
@@ -60,7 +60,7 @@ func (b NBT) Float32(name string) (ret float32, ok bool) {
 }
 
 func (b NBT) Float64(name string) (ret float64, ok bool) {
-	var val interface{}
+	var val any
 	if val, ok = b[name]; ok {
 		ret, ok = val.(float64)
 	}
@@ -68,7 +68,7 @@ func (b NBT) Float64(name string) (ret float64, ok bool) {
 }
 
 func (b NBT) ByteArray(name string) (ret []byte, ok bool) {
-	var val interface{}
+	var val any
 	if val, ok = b[name]; ok {
 		ret, ok = val.([]byte)
 	}
@@ -76,7 +76,7 @@ func (b NBT) ByteArray(name string) (ret []byte, ok bool) {
 }
 
 func (b NBT) String(name string) (ret string, ok bool) {
-	var val interface{}
+	var val any
 	if val, ok = b[name]; ok {
 		ret, ok = val.(string)
 	}
@@ -84,9 +84,9 @@ func (b NBT) String(name string) (ret string, ok bool) {
 }
 
 func (b NBT) NBT(name string) (ret NBT, ok bool) {
-	var val interface{}
+	var val any
 	if val, ok = b[name]; ok {
-		ret, ok = val.(map[string]interface{})
+		ret, ok = val.(map[string]any)
 		if !ok {
 			ret, ok = val.(NBT)
 		}
@@ -95,17 +95,17 @@ func (b NBT) NBT(name string) (ret NBT, ok bool) {
 }
 
 func (b NBT) List(name string) (ret []NBT, ok bool) {
-	var val interface{}
+	var val any
 	if val, ok = b[name]; ok {
-		var l []interface{}
-		l, ok = val.([]interface{})
+		var l []any
+		l, ok = val.([]any)
 		if !ok {
 			ret, ok = val.([]NBT)
 			return
 		}
 		var n NBT
 		for _, e := range l {
-			n, ok = e.(map[string]interface{})
+			n, ok = e.(map[string]any)
 			if ok {
 				ret = append(ret, n)
 			}
@@ -115,7 +115,7 @@ func (b NBT) List(name string) (ret []NBT, ok bool) {
 }
 
 func (b NBT) Int32Array(name string) (ret []int32, ok bool) {
-	var val interface{}
+	var val any
 	if val, ok = b[name]; ok {
 		ret, ok = val.([]int32)
 	}
@@ -123,7 +123,7 @@ func (b NBT) Int32Array(name string) (ret []int32, ok bool) {
 }
 
 func (b NBT) Int64Array(name string) (ret []int64, ok bool) {
-	var val interface{}
+	var val any
 	if val, ok = b[name]; ok {
 		ret, ok = val.([]int64)
 	}
