@@ -27,7 +27,13 @@ func (m *manager) Wait() {
 	m.activeSubscribers.Wait()
 }
 
-func (m *manager) Subscribe(eventType interface{}, priority int, fn HandlerFunc) (unsubscribe func()) {
+func (m *manager) HasSubscribers(event Event) bool {
+	m.mu.RLock()
+	defer m.mu.RUnlock()
+	return len(m.subscribers[TypeOf(event)]) != 0
+}
+
+func (m *manager) Subscribe(eventType any, priority int, fn HandlerFunc) (unsubscribe func()) {
 	eType := TypeOf(eventType)
 	if eType == nil {
 		return func() {}
