@@ -5,9 +5,10 @@ var Nop Manager = &nopMgr{}
 
 type nopMgr struct{}
 
-func (n *nopMgr) Subscribe(interface{}, int, HandlerFunc) (unsubscribe func()) {
+func (n *nopMgr) Subscribe(any, int, HandlerFunc) (unsubscribe func()) {
 	return func() {}
 }
 func (n *nopMgr) Fire(Event)                         {}
 func (n *nopMgr) FireParallel(Event, ...HandlerFunc) {}
 func (n *nopMgr) Wait()                              {}
+func (n *nopMgr) HasSubscribers(event Event) bool    { return false }
