@@ -70,6 +70,7 @@ func init() {
 		m(0x0F, version.Minecraft_1_14),
 		m(0x10, version.Minecraft_1_16),
 		m(0x0F, version.Minecraft_1_17),
+		m(0x11, version.Minecraft_1_19),
 	)
 	Play.ServerBound.Register(&plugin.Message{},
 		m(0x17, version.Minecraft_1_7_2),
@@ -79,6 +80,7 @@ func init() {
 		m(0x0A, version.Minecraft_1_13),
 		m(0x0B, version.Minecraft_1_14),
 		m(0x0A, version.Minecraft_1_17),
+		m(0x0C, version.Minecraft_1_19),
 	)
 	Play.ServerBound.Register(&p.ClientSettings{},
 		m(0x15, version.Minecraft_1_7_2),
@@ -86,13 +88,20 @@ func init() {
 		m(0x05, version.Minecraft_1_12),
 		m(0x04, version.Minecraft_1_12_1),
 		m(0x05, version.Minecraft_1_14),
+		m(0x07, version.Minecraft_1_19),
 	)
-	Play.ServerBound.Register(&p.Chat{},
+	Play.ServerBound.Register(&p.LegacyChat{},
 		m(0x01, version.Minecraft_1_7_2),
 		m(0x02, version.Minecraft_1_9),
 		m(0x03, version.Minecraft_1_12),
 		m(0x02, version.Minecraft_1_12_1),
-		m(0x03, version.Minecraft_1_14),
+		ml(0x03, version.Minecraft_1_14, version.Minecraft_1_18_2),
+	)
+	Play.ServerBound.Register(&p.PlayerCommand{},
+		m(0x03, version.Minecraft_1_19),
+	)
+	Play.ServerBound.Register(&p.PlayerChat{},
+		m(0x04, version.Minecraft_1_19),
 	)
 	Play.ServerBound.Register(&p.TabCompleteRequest{},
 		m(0x14, version.Minecraft_1_7_2),
@@ -101,9 +110,19 @@ func init() {
 		m(0x01, version.Minecraft_1_12_1),
 		m(0x05, version.Minecraft_1_13),
 		m(0x06, version.Minecraft_1_14),
+		m(0x08, version.Minecraft_1_19),
+	)
+	Play.ServerBound.Register(&p.ResourcePackResponse{},
+		m(0x19, version.Minecraft_1_8),
+		m(0x16, version.Minecraft_1_9),
+		m(0x18, version.Minecraft_1_12),
+		m(0x1D, version.Minecraft_1_13),
+		m(0x1F, version.Minecraft_1_14),
+		m(0x20, version.Minecraft_1_16),
+		m(0x21, version.Minecraft_1_16_2),
+		m(0x23, version.Minecraft_1_19),
 	)
 	// coming soon...
-	// ResourcePackResponse
 
 	Play.ClientBound.Register(&p.KeepAlive{},
 		m(0x00, version.Minecraft_1_7_2),
@@ -114,6 +133,7 @@ func init() {
 		m(0x20, version.Minecraft_1_16),
 		m(0x1F, version.Minecraft_1_16_2),
 		m(0x21, version.Minecraft_1_17),
+		m(0x1E, version.Minecraft_1_19),
 	)
 	Play.ClientBound.Register(&p.JoinGame{},
 		m(0x01, version.Minecraft_1_7_2),
@@ -124,6 +144,7 @@ func init() {
 		m(0x25, version.Minecraft_1_16),
 		m(0x24, version.Minecraft_1_16_2),
 		m(0x26, version.Minecraft_1_17),
+		m(0x23, version.Minecraft_1_19),
 	)
 	Play.ClientBound.Register(&p.Respawn{},
 		m(0x07, version.Minecraft_1_7_2),
@@ -136,6 +157,7 @@ func init() {
 		m(0x3A, version.Minecraft_1_16),
 		m(0x39, version.Minecraft_1_16_2),
 		m(0x3D, version.Minecraft_1_17),
+		m(0x3B, version.Minecraft_1_19),
 	)
 	Play.ClientBound.Register(&p.Disconnect{},
 		m(0x40, version.Minecraft_1_7_2),
@@ -146,14 +168,15 @@ func init() {
 		m(0x1A, version.Minecraft_1_16),
 		m(0x19, version.Minecraft_1_16_2),
 		m(0x1A, version.Minecraft_1_17),
+		m(0x17, version.Minecraft_1_19),
 	)
-	Play.ClientBound.Register(&p.Chat{},
+	Play.ClientBound.Register(&p.LegacyChat{},
 		m(0x02, version.Minecraft_1_7_2),
 		m(0x0F, version.Minecraft_1_9),
 		m(0x0E, version.Minecraft_1_13),
 		m(0x0F, version.Minecraft_1_15),
 		m(0x0E, version.Minecraft_1_16),
-		m(0x0F, version.Minecraft_1_17),
+		ml(0x0F, version.Minecraft_1_17, version.Minecraft_1_18_2),
 	)
 	Play.ClientBound.Register(&p.HeaderAndFooter{},
 		m(0x47, version.Minecraft_1_8),
@@ -167,6 +190,7 @@ func init() {
 		m(0x53, version.Minecraft_1_16),
 		m(0x5E, version.Minecraft_1_17),
 		m(0x5F, version.Minecraft_1_18),
+		m(0x60, version.Minecraft_1_19),
 	)
 	Play.ClientBound.Register(&p.PlayerListItem{},
 		m(0x38, version.Minecraft_1_7_2),
@@ -178,6 +202,7 @@ func init() {
 		m(0x33, version.Minecraft_1_16),
 		m(0x32, version.Minecraft_1_16_2),
 		m(0x36, version.Minecraft_1_17),
+		m(0x34, version.Minecraft_1_19),
 	)
 	Play.ClientBound.Register(&title.Legacy{},
 		m(0x45, version.Minecraft_1_8),
@@ -199,6 +224,7 @@ func init() {
 	)
 	Play.ClientBound.Register(&title.Actionbar{},
 		m(0x41, version.Minecraft_1_17),
+		m(0x40, version.Minecraft_1_19),
 	)
 	Play.ClientBound.Register(&title.Times{},
 		m(0x5A, version.Minecraft_1_17),
@@ -206,6 +232,7 @@ func init() {
 	)
 	Play.ClientBound.Register(&title.Clear{},
 		m(0x10, version.Minecraft_1_17),
+		m(0x0D, version.Minecraft_1_19),
 	)
 	Play.ClientBound.Register(&plugin.Message{},
 		m(0x3F, version.Minecraft_1_7_2),
@@ -216,6 +243,7 @@ func init() {
 		m(0x18, version.Minecraft_1_16),
 		m(0x17, version.Minecraft_1_16_2),
 		m(0x18, version.Minecraft_1_17),
+		m(0x15, version.Minecraft_1_19),
 	)
 	Play.ClientBound.Register(&p.ResourcePackRequest{},
 		m(0x48, version.Minecraft_1_8),
@@ -228,6 +256,7 @@ func init() {
 		m(0x39, version.Minecraft_1_16),
 		m(0x38, version.Minecraft_1_16_2),
 		m(0x3C, version.Minecraft_1_17),
+		m(0x3A, version.Minecraft_1_19),
 	)
 	Play.ClientBound.Register(&p.TabCompleteResponse{},
 		m(0x3A, version.Minecraft_1_7_2),
@@ -237,6 +266,7 @@ func init() {
 		m(0x10, version.Minecraft_1_16),
 		m(0x0F, version.Minecraft_1_16_2),
 		m(0x11, version.Minecraft_1_17),
+		m(0x0E, version.Minecraft_1_19),
 	)
 	Play.ClientBound.Register(&p.AvailableCommands{},
 		m(0x11, version.Minecraft_1_13),
@@ -244,6 +274,13 @@ func init() {
 		m(0x11, version.Minecraft_1_16),
 		m(0x10, version.Minecraft_1_16_2),
 		m(0x12, version.Minecraft_1_17),
+		m(0x0F, version.Minecraft_1_19),
+	)
+	Play.ClientBound.Register(&p.SystemChat{},
+		m(0x5F, version.Minecraft_1_19),
+	)
+	Play.ClientBound.Register(&p.ServerData{},
+		m(0x3F, version.Minecraft_1_19),
 	)
 	// coming soon...
 	// BossBar
