@@ -1,7 +1,6 @@
 package proxy
 
 import (
-	"context"
 	"crypto/rand"
 	"encoding/binary"
 	"encoding/hex"
@@ -12,9 +11,12 @@ import (
 	"sync"
 	"time"
 
+	"github.com/gammazero/deque"
 	"github.com/go-logr/logr"
 	"go.minekube.com/common/minecraft/component"
 	"go.minekube.com/common/minecraft/component/codec/legacy"
+	"go.minekube.com/gate/pkg/edition/java/proxy/crypto"
+	"go.minekube.com/gate/pkg/edition/java/proxy/tablist"
 	"go.uber.org/atomic"
 
 	"go.minekube.com/gate/pkg/command"
@@ -38,6 +40,7 @@ type Player interface {
 	command.Source
 	message.ChannelMessageSource
 	message.ChannelMessageSink
+	crypto.KeyIdentifiable
 
 	ID() uuid.UUID    // The Minecraft ID of the player.
 	Username() string // The username of the player.
@@ -55,19 +58,13 @@ type Player interface {
 	// SpoofChatInput sends chats input onto the player's current server as if
 	// they typed it into the client chat box.
 	SpoofChatInput(input string) error
-	// SendResourcePack sends the specified resource pack from url to the user. If at all possible,
-	// send the resource pack with a sha1 hash using SendResourcePackWithHash. To monitor the status
-	// of the sent resource pack, subscribe to PlayerResourcePackStatusEvent.
-	SendResourcePack(url string) error
-	// SendResourcePackWithHash sends the specified resource pack from url to the user,
-	// using the specified 20-byte SHA-1 hash of the resource pack file. To monitor the
-	// status of the sent resource pack, subscribe to PlayerResourcePackStatusEvent.
-	SendResourcePackWithHash(url string, sha1Hash []byte) error
+	// SendResourcePack sends the specified resource pack from url to the user.
+	// If at all possible, specify an 20-byte SHA-1 hash of the resource pack file.
+	// To monitor the status of the sent resource pack, subscribe to PlayerResourcePackStatusEvent.
+	SendResourcePack(info ResourcePackInfo) error
 	// SendActionBar sends an action bar to the player.
 	SendActionBar(msg component.Component) error
-	// SendMessageWith sends a chat message with optional modifications.
-	SendMessageWith(msg component.Component, opts ...MessageOption) error
-	player.TabList
+	TabList() tablist.TabList // Returns the player's tab list.
 	// TODO add title and more
 }
 
@@ -79,6 +76,7 @@ type connectedPlayer struct {
 	profile     *profile.GameProfile
 	ping        atomic.Duration
 	permFunc    permission.Func
+	playerKey   crypto.IdentifiedKey // 1.19+
 
 	// This field is true if this connection is being disconnected
 	// due to another connection logging in with the same GameProfile.
@@ -87,14 +85,17 @@ type connectedPlayer struct {
 	pluginChannelsMu sync.RWMutex // Protects following field
 	pluginChannels   sets.String  // Known plugin channels
 
-	*tabList // Player's tab list
+	tabList tablist.TabList // Player's tab list
 
-	mu               sync.RWMutex // Protects following fields
-	connectedServer_ *serverConnection
-	connInFlight     *serverConnection
-	settings         player.Settings
-	modInfo          *modinfo.ModInfo
-	connPhase        clientConnectionPhase
+	mu                       sync.RWMutex // Protects following fields
+	connectedServer_         *serverConnection
+	connInFlight             *serverConnection
+	settings                 player.Settings
+	modInfo                  *modinfo.ModInfo
+	connPhase                clientConnectionPhase
+	outstandingResourcePacks deque.Deque[*ResourcePackInfo]
+	previousResourceResponse *bool
+	pendingResourcePack      *ResourcePackInfo
 
 	serversToTry []string // names of servers to try if we got disconnected from previous
 	tryIndex     int
@@ -107,9 +108,18 @@ func newConnectedPlayer(
 	profile *profile.GameProfile,
 	virtualHost net.Addr,
 	onlineMode bool,
+	playerKey crypto.IdentifiedKey, // nil-able
 ) *connectedPlayer {
 	ping := atomic.Duration{}
 	ping.Store(-1)
+
+	var tabList tablist.TabList
+	if conn.protocol.GreaterEqual(version.Minecraft_1_8) {
+		tabList = tablist.New(conn, conn.protocol, &tabListPlayerKeyStore{p: conn.proxy})
+	} else {
+		tabList = tablist.NewLegacy(conn, conn.protocol)
+	}
+
 	return &connectedPlayer{
 		minecraftConn: conn,
 		log: conn.log.WithName("player").WithValues(
@@ -120,10 +130,24 @@ func newConnectedPlayer(
 		pluginChannels: sets.NewString(), // Should we limit the size to 1024 channels?
 		connPhase:      conn.Type().initialClientPhase(),
 		ping:           ping,
-		tabList:        newTabList(conn),
+		tabList:        tabList,
 		permFunc:       func(string) permission.TriState { return permission.Undefined },
+		playerKey:      playerKey,
 	}
 }
+
+type tabListPlayerKeyStore struct{ p *Proxy }
+
+var _ tablist.PlayerKey = (*tabListPlayerKeyStore)(nil)
+
+func (t *tabListPlayerKeyStore) PlayerKey(playerID uuid.UUID) crypto.IdentifiedKey {
+	if p := t.p.player(playerID); p != nil {
+		return p.playerKey
+	}
+	return nil
+}
+
+func (p *connectedPlayer) IdentifiedKey() crypto.IdentifiedKey { return p.playerKey }
 
 func (p *connectedPlayer) connectionInFlight() *serverConnection {
 	p.mu.RLock()
@@ -157,6 +181,8 @@ func (p *connectedPlayer) GameProfile() profile.GameProfile {
 	return *p.profile
 }
 
+func (p *connectedPlayer) TabList() tablist.TabList { return p.tabList }
+
 var (
 	ErrNoBackendConnection = errors.New("player has no backend server connection yet")
 	ErrTooLongChatMessage  = errors.New("server bound chat message can not exceed 256 characters")
@@ -171,10 +197,8 @@ func (p *connectedPlayer) SpoofChatInput(input string) error {
 	if !ok {
 		return ErrNoBackendConnection
 	}
-	return serverMc.WritePacket(&packet.Chat{
-		Message: input,
-		Type:    packet.ChatMessageType,
-	})
+	write := packet.NewChatBuilder(p.Protocol()).AsPlayer(p.ID()).Message(input).ToServer()
+	return serverMc.WritePacket(write)
 }
 
 func (p *connectedPlayer) ensureBackendConnection() (*minecraftConn, bool) {
@@ -192,21 +216,169 @@ func (p *connectedPlayer) ensureBackendConnection() (*minecraftConn, bool) {
 	return serverMc, true
 }
 
-func (p *connectedPlayer) SendResourcePack(url string) error {
-	return p.WritePacket(&packet.ResourcePackRequest{
-		Url:  url,
-		Hash: "",
-	})
+func (p *connectedPlayer) SendResourcePack(info ResourcePackInfo) error {
+	if !p.Protocol().GreaterEqual(version.Minecraft_1_8) {
+		return nil
+	}
+	return p.queueResourcePack(info)
 }
 
-func (p *connectedPlayer) SendResourcePackWithHash(url string, sha1Hash []byte) error {
-	if len(sha1Hash) != 20 {
-		return errors.New("hash length must be 20")
+// ResourcePackInfo is resource-pack options for Player.SendResourcePack.
+type ResourcePackInfo struct {
+	// The download link the resource-pack can be found at.
+	URL string
+	// The SHA-1 hash of the provided resource pack.
+	//
+	// Note: It is recommended to always set this hash.
+	// If this hash is not set/not present then the client will always download
+	// the resource pack even if it may still be cached. By having this hash present,
+	// the client will check first whether a resource pack by this hash is cached
+	// before downloading.
+	Hash []byte
+	// Whether the acceptance of the resource-pack is enforced.
+	//
+	// Sets the resource-pack as required to play on the network.
+	// This feature was introduced in 1.17.
+	// Setting this to true has one of two effects:
+	// If the client is on 1.17 or newer:
+	//  - The resource-pack prompt will display without a decline button
+	//  - Accept or disconnect are the only available options but players may still press escape.
+	//  - Forces the resource-pack offer prompt to display even if the player has
+	//    previously declined or disabled resource packs
+	//  - The player will be disconnected from the network if they close/skip the prompt.
+	// If the client is on a version older than 1.17:
+	//   - If the player accepts the resource pack or has previously accepted a resource-pack
+	//     then nothing else will happen.
+	//   - If the player declines the resource pack or has previously declined a resource-pack
+	//     the player will be disconnected from the network
+	ShouldForce bool
+	// The optional message that is displayed on the resource-pack prompt.
+	// This is only displayed if the client version is 1.17 or newer.
+	Prompt component.Component
+	Origin ResourcePackOrigin // The origin of the resource-pack.
+}
+
+// ResourcePackOrigin represents the origin of the resource-pack.
+type ResourcePackOrigin byte
+
+const (
+	DownstreamServerResourcePackOrigin ResourcePackOrigin = iota
+	PluginOnProxyResourcePackOrigin
+)
+
+func (p *connectedPlayer) queueResourcePack(info ResourcePackInfo) error {
+	if info.URL == "" {
+		return errors.New("missing resource-pack url")
 	}
-	return p.WritePacket(&packet.ResourcePackRequest{
-		Url:  url,
-		Hash: hex.EncodeToString(sha1Hash),
-	})
+	if len(info.Hash) > 0 && len(info.Hash) != 20 {
+		return errors.New("resource-pack hash length must be 20")
+	}
+	p.mu.Lock()
+	p.outstandingResourcePacks.PushBack(&info)
+	size := p.outstandingResourcePacks.Len()
+	p.mu.Unlock()
+	if size == 1 {
+		return p.tickResourcePackQueue()
+	}
+	return nil
+}
+
+func (p *connectedPlayer) tickResourcePackQueue() error {
+	p.mu.RLock()
+	queued := p.outstandingResourcePacks.Front()
+	previousResourceResponse := p.previousResourceResponse
+	p.mu.RUnlock()
+
+	// Check if the player declined a resource pack once already
+	if previousResourceResponse != nil && !*previousResourceResponse {
+		// If that happened we can flush the queue right away.
+		// Unless its 1.17+ and forced it will come back denied anyway
+		for {
+			p.mu.Lock()
+			if p.outstandingResourcePacks.Len() != 0 {
+				p.mu.Unlock()
+				break
+			}
+			queued = p.outstandingResourcePacks.Front()
+			p.mu.Unlock()
+			if queued.ShouldForce && p.Protocol().GreaterEqual(version.Minecraft_1_17) {
+				break
+			}
+			_ = p.onResourcePackResponse(packet.DeclinedResourcePackResponseStatus)
+			queued = nil
+		}
+		if queued == nil {
+			// Exit as the queue was cleared
+			return nil
+		}
+	}
+
+	req := &packet.ResourcePackRequest{
+		URL:      queued.URL,
+		Required: queued.ShouldForce,
+		Prompt:   queued.Prompt,
+	}
+	if len(queued.Hash) != 0 {
+		req.Hash = hex.EncodeToString(queued.Hash)
+	}
+	return p.WritePacket(req)
+}
+
+// Processes a client response to a sent resource-pack.
+func (p *connectedPlayer) onResourcePackResponse(status ResourcePackResponseStatus) bool {
+	peek := status == AcceptedResourcePackResponseStatus
+
+	p.mu.Lock()
+	if p.outstandingResourcePacks.Len() == 0 {
+		p.mu.Unlock()
+		return false
+	}
+
+	var queued *ResourcePackInfo
+	if peek {
+		queued = p.outstandingResourcePacks.Front()
+	} else {
+		queued = p.outstandingResourcePacks.PopFront()
+	}
+	p.mu.Unlock()
+
+	e := &PlayerResourcePackStatusEvent{
+		player:        p,
+		status:        status,
+		packInfo:      *queued,
+		overwriteKick: false,
+	}
+	p.proxy.Event().Fire(e)
+
+	if e.Status() == DeclinedResourcePackResponseStatus &&
+		e.PackInfo().ShouldForce &&
+		(!e.OverwriteKick() || e.Player().Protocol().GreaterEqual(version.Minecraft_1_17)) {
+		e.Player().Disconnect(&component.Translation{
+			Key: "multiplayer.requiredTexturePrompt.disconnect",
+		})
+	}
+
+	p.mu.Lock()
+	switch status {
+	case AcceptedResourcePackResponseStatus:
+		b := true
+		p.previousResourceResponse = &b
+		p.pendingResourcePack = queued
+	case DeclinedResourcePackResponseStatus:
+		b := false
+		p.previousResourceResponse = &b
+	case SuccessfulResourcePackResponseStatus:
+		p.previousResourceResponse = nil
+		p.pendingResourcePack = nil
+	case FailedDownloadResourcePackResponseStatus:
+		p.pendingResourcePack = nil
+	}
+	p.mu.Unlock()
+
+	if !peek {
+		_ = p.tickResourcePackQueue()
+	}
+	return queued != nil && queued.Origin != DownstreamServerResourcePackOrigin
 }
 
 func (p *connectedPlayer) VirtualHost() net.Addr {
@@ -217,40 +389,48 @@ func (p *connectedPlayer) Active() bool {
 	return !p.minecraftConn.Closed()
 }
 
-// MessageOption is an option for Player.SendMessageWith.
-type MessageOption func(c *packet.Chat)
-
-// MessageWithSender modifies the sender identity of the chat message.
-func MessageWithSender(id uuid.UUID) MessageOption {
-	return func(c *packet.Chat) { c.Sender = id }
+// WithMessageSender modifies the sender identity of the chat message.
+func WithMessageSender(id uuid.UUID) command.MessageOption {
+	return messageApplyOption(func(o any) {
+		if b, ok := o.(*packet.ChatBuilder); ok {
+			b.AsPlayer(id)
+		}
+	})
 }
 
 // MessageType is a chat message type.
-type MessageType uint8
+type MessageType = packet.MessageType
 
 // Chat message types.
 const (
-	// ChatMessageType is a standard chat message.
-	ChatMessageType MessageType = iota
+	// ChatMessageType is a standard chat message and
+	// lets the chat message appear in the client's HUD.
+	// These messages can be filtered out by the client's settings.
+	ChatMessageType = packet.ChatMessageType
 	// SystemMessageType is a system chat message.
 	// e.g. client is willing to accept messages from commands,
 	// but does not want general chat from other players.
-	SystemMessageType
+	// It lets the chat message appear in the client's HUD and can't be dismissed.
+	SystemMessageType = packet.SystemMessageType
+	// GameInfoMessageType lets the chat message appear above the player's main HUD.
+	// This text format doesn't support many component features, such as hover events.
+	GameInfoMessageType = packet.GameInfoMessageType
 )
 
-// MessageWithType modifies chat message type.
-func MessageWithType(t MessageType) MessageOption {
-	return func(c *packet.Chat) {
-		if t == SystemMessageType {
-			c.Type = packet.SystemMessageType
-		} else {
-			c.Type = packet.ChatMessageType
+// WithMessageType modifies chat message type.
+func WithMessageType(t MessageType) command.MessageOption {
+	return messageApplyOption(func(o any) {
+		if b, ok := o.(*packet.ChatBuilder); ok {
+			b.Type(t)
 		}
-	}
+	})
 }
 
-func (p *connectedPlayer) SendMessage(msg component.Component) error { return p.SendMessageWith(msg) }
-func (p *connectedPlayer) SendMessageWith(msg component.Component, opts ...MessageOption) error {
+type messageApplyOption func(o any)
+
+func (a messageApplyOption) Apply(o any) { a(o) }
+
+func (p *connectedPlayer) SendMessage(msg component.Component, opts ...command.MessageOption) error {
 	if msg == nil {
 		return nil // skip nil message
 	}
@@ -258,15 +438,11 @@ func (p *connectedPlayer) SendMessageWith(msg component.Component, opts ...Messa
 	if err := util.JsonCodec(p.Protocol()).Marshal(m, msg); err != nil {
 		return err
 	}
-	chat := &packet.Chat{
-		Message: m.String(),
-		Type:    packet.ChatMessageType,
-		Sender:  uuid.Nil,
-	}
+	chat := packet.NewChatBuilder(p.Protocol()).Component(msg).AsPlayer(p.ID()).Type(ChatMessageType)
 	for _, o := range opts {
-		o(chat)
+		o.Apply(chat)
 	}
-	return p.WritePacket(chat)
+	return p.WritePacket(chat.ToClient())
 }
 
 var legacyJsonCodec = &legacy.Legacy{}
@@ -298,7 +474,7 @@ func (p *connectedPlayer) SendActionBar(msg component.Component) error {
 	if err != nil {
 		return err
 	}
-	return p.WritePacket(&packet.Chat{
+	return p.WritePacket(&packet.LegacyChat{ // TODO 1.19 chat
 		Message: string(m),
 		Type:    packet.GameInfoMessageType,
 		Sender:  uuid.Nil,
@@ -491,10 +667,6 @@ func (p *connectedPlayer) setSettings(settings *packet.ClientSettings) {
 	})
 }
 
-func (p *connectedPlayer) Closed() <-chan struct{} {
-	return p.minecraftConn.closed
-}
-
 // Settings returns the players client settings.
 // If not known already, returns player.DefaultSettings.
 func (p *connectedPlayer) Settings() player.Settings {
@@ -504,21 +676,6 @@ func (p *connectedPlayer) Settings() player.Settings {
 		return p.settings
 	}
 	return player.DefaultSettings
-}
-
-// returns a new player context that is canceled when:
-//  - connection disconnects
-//  - parent was canceled
-func (c *minecraftConn) newContext(parent context.Context) (ctx context.Context, cancel func()) {
-	ctx, cancel = context.WithCancel(parent)
-	go func() {
-		select {
-		case <-ctx.Done():
-		case <-c.closed: // TODO use context.Context all along so we don't need to start a new goroutine
-			cancel()
-		}
-	}()
-	return ctx, cancel
 }
 
 func randomUint64() uint64 {
